The beginning of... uh, another beginning
WE GO AGAIN.
Started over from scratch. Some systems will be recycled from previous (newer) HUD projects (such as the improved buff tracker from NobleHUD, or the animation system used in several of my mods), but most content will be made specially for this new branch of KineticHUD.
The core tenets of KineticHUD's design will remain, with a few added:
* Customization (Allow placement and scaling of all main hud elements)
* Compatibility (Don't break other mods, don't be broken by other mods; accommodate other mods with compatibility workarounds when necessary)
* Information (add information or move where it's visible, but don't remove player access to it)
* Modularity (Allow individually disabling main HUD elements separately from one another)
* Accessibility (Try to give the same or as similar as possible gameplay experience to all players, accommodating for things like colorblindness where appropriate)

...but the intent behind this new branch is to take another shot at khud with the new experience and better coding practices I've learned since initially creating khud, two years prior to the creation of this branch.
BeardLib is planned to be required for this new HUD mod.
This mod is also still partly in the design phase, but has sufficiently progressed for me to be able to begin the implementation phase.
The buff data and buff hooks are retained from NobleHUD, but are not currently reintegrated into this version of KineticHUD.